+++ demo/CypressLayout.tsx
@@ -22,6 +22,7 @@ import {
   MultiSelect,
   Notifications,
   NumericRange,
+  NumericTextBox,
   Pagination,
   Password,
   ProgressBar,
@@ -113,6 +114,9 @@ export const CypressLayout = () => (
         </Route>
         <Route path="/cypress/numericrange">
           <NumericRange />
+        </Route>
+        <Route path="/cypress/numerictextbox">
+          <NumericTextBox />
         </Route>
         <Route path="/cypress/pagination">
           <Pagination />