+++ korus-ui/src/DateTimeInput/hooks.ts
@@ -6,7 +6,7 @@ import { Span } from '../../components/Span';
 import { useElement } from '../../utils';
 import { VIEW_TYPES } from '../Calendar/constants';
 import { MaskedInputBase } from '../MaskedInputBase';
-import { setDate, setViewDate } from './actions';
+import { setDate, setPrevDate, setViewDate } from './actions';
 import {
   stringToDate,
 } from './helpers';
@@ -40,6 +40,8 @@ export const useDateTimeInputEffects = ({
   React.useEffect(() => {
     // synchronizing the calendar display with value
     if (dateState && conditions.isValueInRange) dispatch(setViewDate(dateState));
+    // memorizing date state
+    if (dateState) dispatch(setPrevDate(dateState));
     // if value is empty reset calendar to today date
     if (dateState === null) {
       const today = new Date();
@@ -49,7 +51,7 @@ export const useDateTimeInputEffects = ({
   }, [conditions.isValueInRange, dispatch, dateState, min, max]);
 
   React.useEffect(() => {
-    // если в value пустая строка - нужно обнулить date для валидации
+    // if the value is an empty string, the date should be reset for validation
     if (isDate(valueProp) || isNil(valueProp) || isFocused) return;
 
     if (valueProp.length === 0) {
@@ -57,7 +59,7 @@ export const useDateTimeInputEffects = ({
     }
 
     const newDate = stringToDate(valueProp, format);
-    // если в инпуте валидная дата - записываем в date, иначе - запиываем null
+    // if the input contains a valid date set this date to state, otherwise set null
     if (newDate && newDate.getDate()) dispatch(setDate(newDate));
     else dispatch(setDate(null));
   }, [dispatch, format, isFocused, max, min, valueProp]);
@@ -68,11 +70,11 @@ export const useDateTimeInputState = (props: DateTimeInputProps): [DateTimeInput
     value: valueProp, format, min, max, isOpen,
   } = props;
 
-  // если сегодняшняя дата за пределами min/max - открываем календарь с датой min или max
+  // if today's date is outside of min/max open the calendar with the date min or max
   const todayIsMin = (min && new Date() < min) ? min : null;
 
   const todayIsMax = (max && new Date() > max) ? max : null;
-  // сегодня, время берется равно 00:00
+  // today, time is 00:00
   const today = new Date();
 
   const stringValue = isDate(valueProp) || isNil(valueProp) ? '' : valueProp;
@@ -82,6 +84,7 @@ export const useDateTimeInputState = (props: DateTimeInputProps): [DateTimeInput
     isFocused: false,
     isOpen: false,
     isValid: true,
+    prevDate: null,
     value: '',
     viewDate: todayIsMin
     || todayIsMax
@@ -128,4 +131,15 @@ export const useCustomElements = (props: DateTimeInputProps, state: DateTimeInpu
     Input,
     Wrapper,
   };
+};
+
+export const useClearMask = (): [string | undefined, () => () => void] => {
+  const [maskedInputValue, setMaskedInputValue] = React.useState<undefined | string>(undefined);
+
+  const clearMaskValue = () => {
+    setMaskedInputValue(undefined);
+    return () => setMaskedInputValue('');
+  };
+
+  return [maskedInputValue, clearMaskValue];
 };