+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Tour: Fixed scroll behaviour when active element is inside modal window
 - TimeRange: Fixed behaviour on Tab click
 - DateRange: Fixed time for date values
+- TimeRange, TimePicker: Added max limits for hours, minutes seconds
+- Changed file structure in README
 
 
 ## [0.37.0] - 2020-12-21